Move retry policy tests into separate class
